requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
requests
psycopg2-binary
boto3
pika